Exterior Scene
Added siheyuan file, then added that to my map tile and rigged it with a sceneswitcher box at the Zhengfang to trasnport you inside. Inside is not rigged with a scene switcher. Also, not everything in Siheyuan has collison due to computer issues and scaling considerations--lastly, the main walkway's collision in the Siheyuan gets you stuck. Not sure why. A ltitle disappointing, but happy with the overall project.